Updated the primary key type for repositories.

+++ src/main/java/dog/ticketlords/TicketlordsBE/repositories/PriceAlertRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import dog.ticketlords.TicketlordsBE.entity.PriceAlert;
 
-public interface PriceAlertRepository extends JpaRepository<PriceAlert, Integer> {
+public interface PriceAlertRepository extends JpaRepository<PriceAlert, Long> {
 }
